requirements.txt
django
djangorestframework
Pillow
requests
django-axes
django-allauth
firebase_admin
psycopg2 # sudo apt install python3-dev libpq-dev for linux
